# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Sample IoT APIs",
    "description" : "These are public domain APIs, currently used for demo purpose.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/thirumalavenkateshc/IOTAPI/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "paths" : {
...
    "/lighting/dimmers/{deviceId}/{value}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "operationId" : "setDimmer",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/dimmers/{deviceId}/{value}/timer/{timeunit}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "description" : "sets a dimmer to a specific value on a timer",
        "operationId" : "setDimmerTimer",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "timeunit",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "units",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "default" : "milliseconds",
            "enum" : [ "seconds", "minutes", "milliseconds" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/switches/{deviceId}" : {
      "get" : {
        "tags" : [ "Z-Wave" ],
        "operationId" : "getSwitchState",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceState"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/switches/{deviceId}/{value}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "operationId" : "setSwitch",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "boolean",
            "enum" : [ true, false ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/switches/{deviceId}/{value}/timer/{minutes}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "description" : "sets a switch to a specific value on a timer",
        "operationId" : "setSwitchTimer",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "boolean",
            "enum" : [ true, false ]
          }
        }, {
          "name" : "minutes",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
...
    "/zones/{zoneId}/quiet" : {
      "get" : {
        "tags" : [ "Zones" ],
        "operationId" : "quietZone",
        "parameters" : [ {
          "name" : "zoneId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "basement", "first-floor", "second-floor" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok"
          }
        },
        "x-swagger-router-controller" : "Zones"
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string",
            "example" : "everything is ok"
          }
        },
        "example" : {
          "code" : 0,
          "message" : "everything is ok"
        }
      },
...
      "DeviceState" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "lastUpdate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "level" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "example" : {
          "level" : 0,
          "lastUpdate" : "2000-01-23T04:56:07.000+00:00",
          "name" : "name",
          "id" : "id"
        }
      },
...
    }
  }
}